deleting setupBrowser, split RBAC into different files for front and API
