+++ e2e-tests/playwright/e2e/plugins/rbac/rbac.spec.ts
@@ -7,18 +7,16 @@ import {
   RoleOverviewPO,
 } from '../../../support/pageObjects/page-obj';
 import { Roles, Response } from '../../../support/pages/rbac';
-import { Common, setupBrowser } from '../../../utils/Common';
+import { Common } from '../../../utils/Common';
 import { UIhelper } from '../../../utils/UIhelper';
 
-test.describe.serial('Test RBAC plugin REST API', () => {
+test.describe.serial('Test RBAC plugin Frontend', () => {
   let common: Common;
   let uiHelper: UIhelper;
   let page: Page;
   let responseHelper: Response;
 
-  test.beforeAll(async ({ browser, baseURL }, testInfo) => {
-    page = (await setupBrowser(browser, testInfo)).page;
-
+  test.beforeAll(async ({ baseURL }) => {
     uiHelper = new UIhelper(page);
     common = new Common(page);
 
@@ -35,222 +33,6 @@ test.describe.serial('Test RBAC plugin REST API', () => {
     const authToken = await getRequest.headerValue('Authorization');
 
     responseHelper = new Response(authToken);
-  });
-
-  test('Test that roles and policies from GET request are what expected', async ({
-    request,
-  }) => {
-    const rolesResponse = await request.get(
-      '/api/permission/roles',
-      responseHelper.getSimpleRequest(),
-    );
-    const policiesResponse = await request.get(
-      '/api/permission/policies',
-      responseHelper.getSimpleRequest(),
-    );
-
-    await responseHelper.checkResponse(
-      await rolesResponse,
-      responseHelper.getExpectedRoles(),
-    );
-    await responseHelper.checkResponse(
-      await policiesResponse,
-      responseHelper.getExpectedPolicies(),
-    );
-  });
-
-  test('Create new role for rhdh-qe, change its name, and deny it from reading catalog entities', async ({
-    request,
-  }) => {
-    const members = ['user:default/rhdh-qe'];
-
-    const firstRole = {
-      memberReferences: members,
-      name: 'role:default/admin',
-    };
-    const rolePostResponse = await request.post(
-      '/api/permission/roles',
-      responseHelper.createRoleRequest(firstRole),
-    );
-    const newRole = {
-      memberReferences: members,
-      name: 'role:default/test',
-    };
-
-    const rolePutResponse = await request.put(
-      '/api/permission/roles/role/default/admin',
-      responseHelper.editRoleRequest(firstRole, newRole),
-    );
-
-    const newPolicy = {
-      entityReference: 'role:default/test',
-      permission: 'catalog-entity',
-      policy: 'read',
-      effect: 'deny',
-    };
-
-    const policyPostResponse = await request.post(
-      '/api/permission/policies',
-      responseHelper.createOrDeletePolicyRequest([newPolicy]),
-    );
-
-    expect(rolePostResponse.ok()).toBeTruthy();
-    expect(rolePutResponse.ok()).toBeTruthy();
-    expect(policyPostResponse.ok()).toBeTruthy();
-  });
-
-  test('Test catalog-entity read is denied', async () => {
-    await uiHelper.openSidebar('Catalog');
-    await uiHelper.selectMuiBox('Kind', 'Component');
-    await uiHelper.verifyTableIsEmpty();
-    await uiHelper.openSidebar('Create...');
-    await uiHelper.verifyText(
-      'No templates found that match your filter. Learn more about',
-      false,
-    );
-  });
-
-  test('Test catalog-entity creation is denied', async () => {
-    expect(
-      await uiHelper.isLinkVisible('Register Existing Component'),
-    ).toBeFalsy();
-  });
-
-  test('PUT catalog-entity read and POST create policies', async ({
-    request,
-  }) => {
-    const oldReadPolicy = [
-      { permission: 'catalog-entity', policy: 'read', effect: 'deny' },
-    ];
-    const newReadPolicy = [
-      { permission: 'catalog-entity', policy: 'read', effect: 'allow' },
-    ];
-    const readPutResponse = await request.put(
-      '/api/permission/policies/role/default/test',
-      responseHelper.editPolicyRequest(oldReadPolicy, newReadPolicy),
-    );
-
-    const createPolicy = [
-      {
-        entityReference: 'role:default/test',
-        permission: 'catalog.entity.create',
-        policy: 'create',
-        effect: 'allow',
-      },
-    ];
-    const createPostResponse = await request.post(
-      '/api/permission/policies',
-      responseHelper.createOrDeletePolicyRequest(createPolicy),
-    );
-
-    expect(readPutResponse.ok()).toBeTruthy();
-    expect(createPostResponse.ok()).toBeTruthy();
-  });
-
-  test('Test catalog-entity read is allowed', async () => {
-    await uiHelper.openSidebar('Catalog');
-    await uiHelper.selectMuiBox('Kind', 'API');
-    await uiHelper.clickLink('Nexus Repo Manager 3');
-  });
-
-  test('Test catalog-entity refresh is denied', async () => {
-    expect(
-      await uiHelper.isBtnVisibleByTitle('Schedule entity refresh'),
-    ).toBeFalsy();
-  });
-
-  test('Test catalog-entity create is allowed', async () => {
-    await uiHelper.openSidebar('Create...');
-    expect(
-      await uiHelper.isLinkVisible('Register Existing Component'),
-    ).toBeTruthy();
-  });
-
-  test('Test bad PUT and PUT catalog-entity update policy', async ({
-    request,
-  }) => {
-    const oldBadPolicy = [
-      { permission: 'catalog-entity', policy: 'refresh', effect: 'allow' },
-    ];
-    const newBadPolicy = [
-      { permission: 'catalog-entity', policy: 'read', effect: 'allow' },
-    ];
-    const badPutResponse = await request.put(
-      '/api/permission/policies/role/default/test',
-      responseHelper.editPolicyRequest(oldBadPolicy, newBadPolicy),
-    );
-
-    const oldGoodPolicy = [
-      {
-        permission: 'catalog.entity.create',
-        policy: 'create',
-        effect: 'allow',
-      },
-    ];
-    const newGoodPolicy = [
-      {
-        permission: 'catalog.entity.refresh',
-        policy: 'update',
-        effect: 'allow',
-      },
-    ];
-    const goodPutResponse = await request.put(
-      '/api/permission/policies/role/default/test',
-      responseHelper.editPolicyRequest(oldGoodPolicy, newGoodPolicy),
-    );
-
-    expect(badPutResponse.ok()).toBeFalsy();
-    expect(goodPutResponse.ok()).toBeTruthy();
-  });
-
-  test('Test that the bad PUT didnt go through and catalog-entities can be read', async () => {
-    await uiHelper.openSidebar('Home');
-    await uiHelper.openSidebar('Create...');
-    expect(
-      await uiHelper.isTextVisible(
-        'No templates found that match your filter. Learn more about',
-      ),
-    ).toBeFalsy();
-  });
-
-  test('Test that the good PUT request went through and catalog-entities can be refreshed', async () => {
-    await uiHelper.openSidebar('Catalog');
-    await uiHelper.selectMuiBox('Kind', 'API');
-    await uiHelper.clickLink('Nexus Repo Manager 3');
-    expect(
-      await uiHelper.isBtnVisibleByTitle('Schedule entity refresh'),
-    ).toBeTruthy();
-  });
-
-  test('Test that the good PUT request went through and catalog-entities cant be created', async () => {
-    await uiHelper.openSidebar('Create...');
-    expect(
-      await uiHelper.isLinkVisible('Register Existing Component'),
-    ).toBeFalsy();
-  });
-
-  test('DELETE catalog-entity update policy', async ({ request }) => {
-    const deletePolicies = [
-      {
-        entityReference: 'role:default/test',
-        permission: 'catalog.entity.refresh',
-        policy: 'update',
-        effect: 'allow',
-      },
-    ];
-    const deleteResponse = await request.delete(
-      '/api/permission/policies/role/default/test',
-      responseHelper.createOrDeletePolicyRequest(deletePolicies),
-    );
-
-    expect(deleteResponse.ok()).toBeTruthy();
-  });
-
-  test('Test catalog-entity refresh is denied after DELETE', async () => {
-    await uiHelper.openSidebar('Catalog');
-    await uiHelper.selectMuiBox('Kind', 'API');
-    await uiHelper.clickLink('Nexus Repo Manager 3');
-    expect(await uiHelper.isBtnVisible('Schedule entity refresh')).toBeFalsy();
   });
 
   test.afterAll(
@@ -279,6 +61,58 @@ test.describe.serial('Test RBAC plugin REST API', () => {
       expect(deleteRole.ok()).toBeTruthy();
     },
   );
+
+  test('Test catalog-entity read is denied', async () => {
+    await uiHelper.openSidebar('Catalog');
+    await uiHelper.selectMuiBox('Kind', 'Component');
+    await uiHelper.verifyTableIsEmpty();
+    await uiHelper.openSidebar('Create...');
+    await uiHelper.verifyText(
+      'No templates found that match your filter. Learn more about',
+      false,
+    );
+  });
+
+  test('Test catalog-entity creation is denied', async () => {
+    expect(
+      await uiHelper.isLinkVisible('Register Existing Component'),
+    ).toBeFalsy();
+  });
+
+  test('Test catalog-entity read is allowed', async () => {
+    await uiHelper.openSidebar('Catalog');
+    await uiHelper.selectMuiBox('Kind', 'API');
+    await uiHelper.clickLink('Nexus Repo Manager 3');
+  });
+
+  test('Test catalog-entity refresh is denied', async () => {
+    expect(
+      await uiHelper.isBtnVisibleByTitle('Schedule entity refresh'),
+    ).toBeFalsy();
+  });
+
+  test('Test catalog-entity create is allowed', async () => {
+    await uiHelper.openSidebar('Create...');
+    expect(
+      await uiHelper.isLinkVisible('Register Existing Component'),
+    ).toBeTruthy();
+  });
+
+  test('Test that the good PUT request went through and catalog-entities can be refreshed', async () => {
+    await uiHelper.openSidebar('Catalog');
+    await uiHelper.selectMuiBox('Kind', 'API');
+    await uiHelper.clickLink('Nexus Repo Manager 3');
+    expect(
+      await uiHelper.isBtnVisibleByTitle('Schedule entity refresh'),
+    ).toBeTruthy();
+  });
+
+  test('Test catalog-entity refresh is denied after DELETE', async () => {
+    await uiHelper.openSidebar('Catalog');
+    await uiHelper.selectMuiBox('Kind', 'API');
+    await uiHelper.clickLink('Nexus Repo Manager 3');
+    expect(await uiHelper.isBtnVisible('Schedule entity refresh')).toBeFalsy();
+  });
 });
 
 test.describe
@@ -287,9 +121,7 @@ test.describe
   let uiHelper: UIhelper;
   let page: Page;
 
-  test.beforeAll(async ({ browser }, testInfo) => {
-    page = (await setupBrowser(browser, testInfo)).page;
-
+  test.beforeAll(async () => {
     uiHelper = new UIhelper(page);
     common = new Common(page);
     await common.loginAsGithubUser();
@@ -352,9 +184,7 @@ test.describe
   let uiHelper: UIhelper;
   let page: Page;
 
-  test.beforeAll(async ({ browser }, testInfo) => {
-    page = (await setupBrowser(browser, testInfo)).page;
-
+  test.beforeAll(async () => {
     uiHelper = new UIhelper(page);
     common = new Common(page);
     await common.loginAsGithubUser(process.env.GH_USER2_ID);
@@ -401,9 +231,7 @@ test.describe.serial('Test RBAC plugin as an admin user', () => {
   let page: Page;
   let rolesHelper: Roles;
 
-  test.beforeAll(async ({ browser }, testInfo) => {
-    page = (await setupBrowser(browser, testInfo)).page;
-
+  test.beforeAll(async () => {
     uiHelper = new UIhelper(page);
     common = new Common(page);
     rolesHelper = new Roles(page);
